requirements.txt
Flask==3.1.1
numpy>=1.26.0,<2.2.0
Pillow==11.2.1
tensorflow==2.19.0
ultralytics==8.3.140
gunicorn==23.0.0
numpy
pandas
tqdm
scikit-learn
matplotlib
seaborn
pillow
torch
torchvision